requirements.txt
streamlit
requests
pandas
python-dotenv
google-generativeai
pypdf
langchain
langchain-google-genai
chromadb